requirements.txt
grpcio~=1.28.1
grpcio-health-checking~=1.28.1
hydro-serving-grpc==2.2.1
docker~=4.2.0
pytest~=5.4.1